requirements.txt
PyQt6>=6.0.0
Pillow>=10.0.0